+++ autodocodec-api-usage/test/Autodocodec/AesonSpec.hs
@@ -104,6 +104,7 @@ spec = do
   aesonCodecSpec @Example
   aesonCodecErrorSpec @Example "example-error-bool-number" "{\"text\": \"hello\", \"bool\": 5}"
   aesonCodecErrorSpec @Example "example-error-fruit-number" "{\"text\": \"hello\", \"bool\": true, \"maybe\": null, \"fruit\": 5}"
+  aesonCodecSpec @Derived
   aesonCodecSpec @Recursive
   aesonCodecSpec @ListsExample
   aesonCodecErrorSpec @Recursive "recursive-error-recurse-string" "{\"recurse\": {\"recurse\": {\"recurse\": \"hello\"}}}"