+++ autodocodec-servant-multipart/default.nix
@@ -4,7 +4,7 @@
 }:
 mkDerivation {
   pname = "autodocodec-servant-multipart";
-  version = "0.0.0.1";
+  version = "0.0.0.2";
   src = ./.;
   libraryHaskellDepends = [
     aeson autodocodec base bytestring servant-multipart