+++ autodocodec-exact/CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
 
+## [0.0.0.1] - 2025-06-20
+
+### Changed
+
+* Lower bound on `autodocodec >=0.5`