+++ autodocodec-servant-multipart/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.0.2] - 2025-06-20
+
+### Changed
+
+* Support for `autodocodec >=0.5`
+
 ## [0.0.0.0] - 2022-07-27
 
 First version